pyproject.toml
[tool.poetry]
name = "ska-mid-cbf-fhs-vcc"
version = "0.0.1"
description = "SKA Mid.CBF FHS VCC"
authors = [
    "Jason Turner <jason.turner@mda.space>",
    "Ben Herriott <ben.herriott@mda.space>",
    "Justin Wamback <justin.wamback@mda.space>"
]
license = "BSD license"
readme = "README.md"
packages = [
    { include = "ska_mid_cbf_fhs_vcc", from = "src" }
]

[tool.poetry.scripts]
Mac200='ska_mid_cbf_fhs_vcc.mac.mac_200_device:main'
PacketValidation='ska_mid_cbf_fhs_vcc.packet_validation.packet_validation_device:main'
B123VccOsppfbChanneliser='ska_mid_cbf_fhs_vcc.b123_vcc_osppfb_channeliser.b123_vcc_osppfb_channeliser_device:main'
FrequencySliceSelection='ska_mid_cbf_fhs_vcc.frequency_slice_selection.frequency_slice_selection_device:main'
WidebandFrequencyShifter='ska_mid_cbf_fhs_vcc.wideband_frequency_shifter.wideband_frequency_shifter_device:main'
WidebandInputBuffer='ska_mid_cbf_fhs_vcc.wideband_input_buffer.wideband_input_buffer_device:main'

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
backoff = "^2.2.1"
typing-extensions = "^4.8.0"
pytango = "^9.5.1"
pyyaml = "^6.0.1"
requests = "2.28.0"
ska-tango-base = "1.0.0"
ska-control-model = "1.0.0"
ska-tango-testing = "0.7.1"
ska-ser-log-transactions = "^0.2.2+ae0f5c92"
ska-telmodel = "1.15.0"
polling2 = "^0.5.0"
dataclasses-json = "^0.6.7"
debugpy = "^1.8.5"

[tool.poetry.group.dev.dependencies]
pytest-pydocstyle = "2.3.2"
pytest-pycodestyle = "2.3.1"
pytest = "8.3.1"
pytest-bdd = "^7.2.0"
pytest-cov = "^5.0.0"
pytest-forked = "^1.6.0"
pytest-json-report = "^1.5.0"
pytest-mock = "^3.14.0"
pytest-xdist = "^3.6.1"
pytest-repeat = "^0.9.3"
pytest-pylint = "^0.21.0"
pylint = "3.2.6"
pylint-junit = "0.3.4"

[tool.poetry.group.docs.dependencies]
Sphinx = "^6"
ska-ser-sphinx-theme = "^0.1.1"
myst-parser = "^2.0.0"
sphinx_collapse = "^0.1.3"
sphinxcontrib-mermaid = "^0.9.2"
black = "^23.9.1"
flake8 = "^6.1.0"
sphinxcontrib-plantuml = "^0.26"

[[tool.poetry.source]]
name = "ska-nexus"
url = "https://artefact.skao.int/repository/pypi-internal/simple"


[[tool.poetry.source]]
name = "PyPI-public"
url = "https://pypi.org/simple"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 79

[build-system]
requires = ["poetry>=1.8", "poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
